w1: replace dt detection by automatic detection
This patch changes the functioning of the detection of w1 devices.
The old way was a comparison between detected w1 and the ones described in
the device tree. Now it will just look for the driver matching the family
id of the w1 detected.

The patch is inspired from Maxime Ripart code.

Signed-off-by: Kory Maincent <kory.maincent@bootlin.com>

---

Change since v2:
- drop the remove action of devicetree match in the W1 EEPROMs drivers.
- move the w1_device_register function to w1-uclass.
- update the w1_device_register funtion to be compatible with automatic
  detection and devicetree. I will not bind the device if it as been found
  in the devicetree.
- update the w1 device driver list to have only two parameters in the
  U_BOOT_W1_DEVICE function.